% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_id_existence.R
\name{check_id_existence}
\alias{check_id_existence}
\title{Check existence of ID columns across two tables}
\usage{
check_id_existence(x, y, by, select_cols, ...)
}
\arguments{
\item{x}{data.frame or tibble containing match id to check for non existence in y}

\item{y}{data.frame or tibble to check for non-existence of match id from x}

\item{by}{character containing match id, or if named different, a named character vector like c("a" = "b")}

\item{select_cols}{character vector of columns to select in the output. Note that during the join, columns with identical names in both data sets will have a suffix of .x or .y added to disambiguate. These need to be added to ensur the correct column is returned.}

\item{...}{other variables passed to dplyr::anti_join}
}
\value{
tibble rows from x without a match in y
}
\description{
This returns rows in x without a match in y. Returning selected columns only. It
is a this wrapper around `dplyr::anti_join`.
}
\examples{
\dontrun{
check_id_existence(x,
                   y,
                   by =  c("Batch_ID" = "batch_id"),
                   select_cols = c("Batch_ID", "iDate", "Farm_ID"))
}
}
\seealso{
`dplyr::anti_join`
}
