% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_to_col_type.R
\docType{data}
\name{class_to_col_type}
\alias{class_to_col_type}
\title{Class to Column Type lookup table}
\format{
## `class_to_col_type`
A data frame with 9 rows and 3 columns:
\describe{
  \item{col_type}{Type of column as described in `readr`}
  \item{col_class}{Class of R object that matches that column type}
  \item{col_abv}{Abbreviation for that column type from `reader`}
  ...
}
}
\usage{
class_to_col_type
}
\description{
A table that links classes to `readr` column types.
Created from csv file of the same name in inst/
}
\seealso{
[reader::cols()]
}
\keyword{datasets}
