% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_logs.R
\name{combine_logs}
\alias{combine_logs}
\title{Combine Validation Logs}
\usage{
combine_logs(existing_log, new_log)
}
\arguments{
\item{existing_log}{tibble existing validation log}

\item{new_log}{tibble newly generated validation log}
}
\value{
tibble appended validation log for upload
}
\description{
Checks for the existence of an existing validation log and appends
new records from the current run.
}
