% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_data.R
\name{correct_data}
\alias{correct_data}
\title{Correct data using validation log}
\usage{
correct_data(validation_log, data, primary_key)
}
\arguments{
\item{validation_log}{tibble a validation log}

\item{data}{tibble the original unclean data}

\item{primary_key}{character the quoted column name for the unique identifier in data}
}
\value{
tibble the semi-clean data set
}
\description{
Takes a validation log and applies the required changes to the data
}
