% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_other_choice_log.R
\name{create_freetext_log}
\alias{create_freetext_log}
\title{create other choice log}
\usage{
create_freetext_log(...)
}
\arguments{
\item{...}{arguments passed to other_choice_log}
}
\description{
use \code{other_choice_log()}
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
