% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_full_id_log.R
\name{create_full_id_log}
\alias{create_full_id_log}
\title{Create the full set of id validation logs}
\usage{
create_full_id_log(dataset_list)
}
\arguments{
\item{dataset_list}{List. A list of datasets where each item has the following
fields semiclean_data, foreign_key, primary_key, and name.}
}
\value{
dataframe id validation log
}
\description{
Takes a list of datasets with shared foreign keys (identifiers that should match
across the datasets) and creates the full set of pairwise id validation logs.
This function is essentially a wrapper for \code{\link[=create_id_log]{create_id_log()}} that computes
all the unique pairs of datasets in a given list then runs each pair through
\code{create_id_log}.
}
\details{
\code{create_id_log} flags duplicate foreign keys and any foreign keys not in both
datasets. The purpose of this log is to ensure links between datasets created
by foreign keys are correct and to help identify any potentially missing data.

Foreign keys are identifiers that link two or more datasets.

Primary keys are identifiers that uniquely identify items in a dataset.
}
\examples{

# dataset 1 and dataset 2 have perfectly matching foreign keys
# and the fields are consistently named.

dataset_item_1 <- list(
semiclean_data = data.frame(primary_key = 1:10,
                            foreign_key = letters[1:10]),
foreign_key = "foreign_key",
primary_key = "primary_key",
name = "dataset_1"
)


dataset_item_2 <- list(
  semiclean_data = data.frame(primary_key = 1:10,
                              foreign_key = letters[1:10]),
  foreign_key = "foreign_key",
  primary_key = "primary_key",
  name = "dataset_2"
)

# dataset 3 uses a different column name for the shared "foreign_key", is missing a,
# and has an additional key k
dataset_item_3 <- list(
  semiclean_data = data.frame(blah = 1:10,
                              bob = letters[2:11]),
  foreign_key = "bob",
  primary_key = "blah",
  name = "dataset_3"
)

# dataset 4 uses yet another name for the shared "foreign_key", has duplicate a's,
# and is missing b

dataset_item_4 <- list(
  semiclean_data = data.frame(blah = 1:10,
                              gary = c(rep(letters[1],2),letters[3:10])
  ),
  foreign_key = "gary",
  primary_key = "blah",
  name = "dataset_4"
)

dataset_list <- list(dataset_item_1,dataset_item_2,dataset_item_3, dataset_item_4)

create_full_id_log(dataset_list)

}
