% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_id_log.R
\name{create_id_log}
\alias{create_id_log}
\title{ID Validation and Verification log}
\usage{
create_id_log(
  semiclean_x,
  semiclean_y,
  by,
  primary_key_x,
  primary_key_y,
  name_x,
  name_y
)
}
\arguments{
\item{semiclean_x}{data.frame or tibble containing match id to check for non existence in y}

\item{semiclean_y}{data.frame or tibble to check for non-existence of match id from x}

\item{by}{character containing match id, or if named different, a named character vector like c("a" = "b")
this function assumes that the named character vector follows the convention of
c("variable in semiclean_x" = "variable in semiclean_y")}

\item{primary_key_x}{character vector of the row identifier for data.frame x}

\item{primary_key_y}{character vector of the row identifier for data.frame y}

\item{name_x}{character of file name for semiclean_x}

\item{name_y}{character of file name for semiclean_y}
}
\value{
tibble formatted as a validation log for human review
}
\description{
This function does two things:
\enumerate{
\item Pairwise comparison of ID (i.e. human, animal, household) columns across two
data sets.
\item Flags duplicate IDs within a data set.
}
}
\details{
This allows for ID cleaning across data sets and sets you up for integration.
}
\examples{
\dontrun{
#testing no differences and no duplicates
semiclean_x<-data.frame("primary_key_x" = 1:10, "hhid" = letters[1:10])
semiclean_y<-data.frame("primary_key_y" = 1:10, "household_id" = letters[1:10])

test_no_mm<-create_id_log(semiclean_x, semiclean_y, by = c("hhid"="household_id"), "primary_key_x",
                         "primary_key_y", name_x = "semiclean_x.csv" , name_y = "semiclean_y.csv")

#should be an empty log

#testing 2 differences
semiclean_x<-data.frame("primary_key_x" = 1:10, "hhid" = letters[1:10])
semiclean_y<-data.frame("primary_key_y" = 1:10, "household_id" = letters[2:11])

test_mm_2<-create_id_log(semiclean_x, semiclean_y, by = c("hhid"="household_id"), "primary_key_x",
                        "primary_key_y", name_x = "semiclean_x.csv" , name_y = "semiclean_y.csv")

#should have a log with 2 rows of IDs, one from y (household_id) and one from x (hhid)
}
}
\seealso{
\code{dplyr::anti_join}
}
