% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_questionnaire_log.R
\name{create_questionnaire_log}
\alias{create_questionnaire_log}
\title{Create Validation Log for Questionnaire data}
\usage{
create_questionnaire_log(data, form_schema, pkey, rule_set, url)
}
\arguments{
\item{data}{data fame  Input data to be validated}

\item{form_schema}{data frame The ODK form schema data}

\item{pkey}{character  A character vector giving the column name of the primary key or unique row identifier in the data}

\item{rule_set}{a rule set of class validator from the validate package}

\item{url}{The ODK submission URL excluding the uuid identifier}
}
\value{
a data frame formatted as a validation log for human review
}
\description{
Create Validation Log for Questionnaire data
}
