% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_rules_from_template.R
\name{create_rules_from_template}
\alias{create_rules_from_template}
\title{Create a "rules" file from a template}
\usage{
create_rules_from_template(name, dir = "R", open = TRUE, showWarnings = FALSE)
}
\arguments{
\item{name}{String. Name of rule set function e.g. create_rules_my_dataset}

\item{dir}{String. Name of directory where file should be created. If it
doesnt exist, a folder will be created.}

\item{open}{Logical. Should the file be opened?}

\item{showWarnings}{Logical. Should dir.create show warnings?}
}
\value{
String. File path of newly created file
}
\description{
Creates a rules file from a template to show general structure of the rule file.
}
\examples{
\dontrun{
# create a ruleset and immediately open it
    create_rules_from_template(name = "create_rules_field_data")
# create a ruleset and don't open it
    create_rules_from_template(name = "create_rules_lab_data", open = FALSE)
# create a ruleset and store it in a different folder
    create_rules_from_template(name = "create_rules_lab_data",
    dir = "/path/to/rulesets" open = FALSE)
    }
}
