% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_validation_log.R
\name{create_validation_log}
\alias{create_validation_log}
\title{Create Validation Log}
\usage{
create_validation_log(data, pkey, rule_set, ...)
}
\arguments{
\item{data}{data fame  Input data to be validated}

\item{pkey}{character  A character vector giving the column name of the primary key or unique row identifier in the data}

\item{rule_set}{a rule set of class validator from the validate package}

\item{...}{other arguments passed to validate::confront}
}
\value{
a data frame formatted as a validation log for human review
}
\description{
Create Validation Log
}
