% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_dropbox.R
\name{download_dropbox}
\alias{download_dropbox}
\title{Download Drop Box}
\usage{
download_dropbox(dropbox_path, dropbox_filename, download_path, ...)
}
\arguments{
\item{dropbox_path}{character The formal folder path on dropbox}

\item{dropbox_filename}{character The formal file name on dropbox}

\item{download_path}{character Local file path to download file to}

\item{...}{other arguments passed to rdrop2::drop_download}
}
\value{
returns file path if successful
}
\description{
Downloads files from dropbox into a given directory
}
