% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_googledrive_files.R
\name{download_googledrive_files}
\alias{download_googledrive_files}
\title{Download Google Drive Files}
\usage{
download_googledrive_files(
  key_path = here::here("./auth/rvf2-workflow-automation-2877f538ddb9.json"),
  drive_path,
  search_pattern,
  MIME_type = NULL,
  out_path
)
}
\arguments{
\item{key_path}{character path to google authentication key}

\item{drive_path}{character The google drive folder path}

\item{search_pattern}{character A search pattern for files in the google drive}

\item{MIME_type}{character Google Drive file type, file extension, or MIME type.}

\item{out_path}{character The local file directory for files to be downloaded to}
}
\value{
vector a character vector of files downloaded
}
\description{
For a given Google Drive folder this function will find and download all files
matching a given pattern.
}
\details{
Note: This relies on the `googledrive::drive_ls()` function which uses a search function
and is not deterministic when recursively searching.
}
