% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_validated_entries.R
\name{drop_validated_entries}
\alias{drop_validated_entries}
\title{Drop Validated Entries in Log}
\usage{
drop_validated_entries(existing_log, new_log)
}
\arguments{
\item{existing_log}{data.frame A log created with the ohcleandat package}

\item{new_log}{data.frame A different log created with the ohcleandat package}
}
\value{
data.frame An ohcleandat log based on the existing log.
}
\description{
Removes any entries that have been validated in the existing log.
Will return records that are in both logs and have no_change = NA
}
\examples{
\dontrun{
 existing_log <- get_dropbox_val_logs(file_name = "log.csv", folder = NULL)

 odk_schema_data <- schema_from_odk_xlsx_template(file_path = "inst/example_odk_schema.xlsx")

 new_log  <- create_free_text_log(response_data = semi_clean_data,
                      form_schema = odk_schema_data,
                      url = "https://odk.xyz.io/#/projects/project-name/submissions",
                      existing_log = existing_log)

 drop_validated_entries(existing_log,new_log)
}

}
