% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dropbox_val_logs.R
\name{get_dropbox_val_logs}
\alias{get_dropbox_val_logs}
\title{Get Dropbox validation logs}
\usage{
get_dropbox_val_logs(
  file_name,
  folder,
  path_name = "DTRA_RVF2/Data/rvf2_github_data/validation_logs"
)
}
\arguments{
\item{file_name}{character file name with extension of the validation log}

\item{folder}{character the folder the log is saved in on drop box}

\item{path_name}{character the default drop box path}
}
\value{
tibble  Validation Log
}
\description{
Downloads validation logs that are stored on dropbox
}
\details{
Ths function will check if the log exists and return NULL if not. Else it will
locally download the file to 'dropbox_validations' directory and read in to the
session.
}
