% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_google_sheets_data.R
\name{get_google_sheets_data}
\alias{get_google_sheets_data}
\title{Get Data from Google Sheets}
\usage{
get_google_sheets_data(ss)
}
\arguments{
\item{ss}{Google sheet id}
}
\value{
List of dataframes from googlesheet
}
\description{
Get Data from Google Sheets
}
