% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_odk_form_schema.R
\name{get_odk_form_schema}
\alias{get_odk_form_schema}
\title{Get ODK Questionnaire Schema Info}
\usage{
get_odk_form_schema(
  url,
  un = Sys.getenv("ODK_USERNAME"),
  pw = Sys.getenv("ODK_PASSWORD")
)
}
\arguments{
\item{url}{character The survey URL}

\item{un}{character The ODK account username}

\item{pw}{character The ODK account password}
}
\value{
data frame of survey responses
}
\description{
This function handles the authentication and pulling of questionnaire
form schema information.
}
