% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obfuscate_gps.R
\name{get_precision}
\alias{get_precision}
\title{Get Precision}
\usage{
get_precision(x, func = c)
}
\arguments{
\item{x}{Numeric. Vector of gps points}

\item{func}{Function. Apply some function to the vector of precisions. Default is c so that
all values are returned}
}
\value{
output of func - likely a vector
}
\description{
Get Precision
}
\examples{

x <- c(1,100,1.11)
get_precision(x,func = min)


}
\author{
Nathan Layman
}
