% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_species_letter.R
\name{get_species_letter}
\alias{get_species_letter}
\title{Get Species Letter}
\usage{
get_species_letter(
  species = c("human", "cattle", "small_mammal", "sheep", "goat")
)
}
\arguments{
\item{species}{character The species identifier. See argument options}
}
\value{
character The hum_anim_id code
}
\description{
This function maps the relationship between animal species and hum_anim_id codes.
This is for use in id_checker()
}
