% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_col_type.R
\name{guess_col_type}
\alias{guess_col_type}
\title{Guess the column type}
\usage{
guess_col_type(data, default_col_abv = "c")
}
\arguments{
\item{data}{data.frame Data who column types you would like to guess}

\item{default_col_abv}{string. Column type abbreviation from [readr::cols()].
Use "g" to guess the column type.}
}
\value{
character vector of column abbreviations
}
\description{
uses column class to set readr column type
}
\examples{
data <- data.frame(time = Sys.time(),
char = "hello", num = 1, log = TRUE,
date = Sys.Date(), list_col = list("hello") )

guess_col_type(data)

## change default value of default column abbreviation

guess_col_type(data, default_col_abv = "g")


}
