% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_checker.R
\name{id_checker}
\alias{id_checker}
\title{ID Checker}
\usage{
id_checker(col, type = c("animal", "hum_anim", "site"), ...)
}
\arguments{
\item{col}{The vector of ID's to be checked}

\item{type}{The ID type, see argument options for allowable settings}

\item{...}{other function arguments passed to get_species_letter}
}
\value{
vector of corrected ID's
}
\description{
General function for checking and correcting ID columns.
}
\details{
In order to use the autobot process for correcting ID columns, a new 'corrected'
column is created by the user using the id_checker() function. It will take an
existing vector of ID's, and an ID type (animal, mosquito, etc) and apply the
bespoke corrections. This can then be consumed by the autobot log.
}
\examples{
\dontrun{
# with a species identifier
    data |> mutate(animal_id_new = id_checker(animal_id, type = "animal", species = "cattle"))
    data |> mutate(farm_id_new = id_checker(farm_id, type = "site"))
    }
}
