% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_report_urls.R
\name{make_report_urls}
\alias{make_report_urls}
\title{Make the URLs for the reports}
\usage{
make_report_urls(aws_deploy_target, pattern = "")
}
\arguments{
\item{aws_deploy_target}{List. Output from aws_s3_upload}

\item{pattern}{String. Regex pattern for matching file paths}
}
\value{
character URL for report
}
\description{
Several HTML reports are emailed via an automated process. To do this a secure
URL is to be generated as a download link. This function is to be used in an
opinionated targets pipeline.
}
\author{
Collin Schwantes
}
