% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_zip_path.R
\name{make_zip_path}
\alias{make_zip_path}
\title{Get make a zip file path}
\usage{
make_zip_path(file_path)
}
\arguments{
\item{file_path}{character.}
}
\value{
character. String where extension is replaced by zip
}
\description{
Take a file path, remove the extension, replace the extension
with .zip
}
\examples{

file_path <- "hello.csv"
make_zip_path(file_path)

file_path_with_dir <- "foo/bar/hello.csv"
make_zip_path(file_path_with_dir)

}
