% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obfuscate_gps.R
\name{obfuscate_gps}
\alias{obfuscate_gps}
\title{Obfuscate GPS}
\usage{
obfuscate_gps(x, precision = 2, fuzz = 0.125)
}
\arguments{
\item{x}{Numeric. Vector of gps points}

\item{precision}{Integer. Number of digits to keep. See \code{round} for more details}

\item{fuzz}{Numeric. Error to introduce to the gps measurements. This is used
to generate the random uniform distribution \code{runif(1,min = -fuzz, max = fuzz)}}
}
\value{
Numeric. A vector of fuzzed and rounded GPS points
}
\description{
This function fuzzes gps points (or any other numeric values) by first adding
error then rounding to a certain number of digits.
}
\examples{

gps_data  <- data.frame(lat = c(1.0001, 10.22223, 4.00588),
  lon = c(2.39595, 4.506930, -60.09999901))

gps_data |>
  # default obfuscation settings correspont to roughly a 27 by 27 km area
  dplyr::mutate(fuzzed_lat = obfuscate_gps(lat) \%>\%
  # can be made more or less precise by changing the number of decimal points
  # included or modifying the amount of fuzz (error) introduced
  dplyr::mutate(fuzzed_lon = obfuscate_gps(lon, precision = 4, fuzz = 0.002))

}
