% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/othertext_lookup.R
\name{othertext_lookup}
\alias{othertext_lookup}
\title{Look-up table for 'Other' questions}
\usage{
othertext_lookup(questionnaire = c("animal_owner"))
}
\arguments{
\item{questionnaire}{The ODK questionnaire. Used to ensure the correct look up table is found.}
}
\value{
tibble
}
\description{
Provides a look up table matching ODK survey questions with their free text
response question.
}
\details{
In many ODK surveys, a multiple choice question can have a response for 'other'
where the respondent can add free text as a response. There is no consistent
link in the response data to match the captured responses and the other free-text
collected. This function provides a manual look up reference so free text responses
can be compared to the original questions in the validation workflow.

This function can be expanded by providing a tibble with two columns: `name` and
`other_name` which maps the question name in ODK to the question name containing
'other' or 'free text'.
}
\examples{
othertext_lookup(questionnaire = c("animal_owner"))


}
