% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_frictionless_metadata.R
\name{prune_datapackage}
\alias{prune_datapackage}
\title{Prune data pacakge}
\usage{
prune_datapackage(my_data_schema, structural_metadata)
}
\arguments{
\item{my_data_schema}{list. schema object from frictionless}

\item{structural_metadata}{dataframe. structural metadata for a dataset}
}
\value{
pruned data_schema -
}
\description{
method to remove properties from the metadata for a dataset in a datapackage
}
