% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_excel_all_sheets.R
\name{read_excel_all_sheets}
\alias{read_excel_all_sheets}
\title{Reads all tabs from an excel workbook}
\usage{
read_excel_all_sheets(file)
}
\arguments{
\item{file}{character File path to an excel file}
}
\value{
list
}
\description{
For a given excel file, this will detect all sheets, and iteratively read
all sheets and place them in a list.
}
