% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropbox_refresh.R
\name{refresh_db_token}
\alias{refresh_db_token}
\title{Dropbox Refresh}
\usage{
refresh_db_token(path = "auth/db_token.rds")
}
\arguments{
\item{path}{to drop box token `.rds` object}
}
\value{
invisible
}
\description{
This is a helper function to trigger the auto-refresh of a dropbox token
}
\details{
This is not an exported function and is only called in download_dropbox() as
a means of refreshing a token. It works by forcing the download of a small sample
file from a fixed file path on EcoHealth drop box.
}
