% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema_from_xlsx.R
\name{ru_odk_case}
\alias{ru_odk_case}
\title{Convert to same case as in RuODK}
\usage{
ru_odk_case(x)
}
\arguments{
\item{x}{Character. Some character vector.}
}
\value{
Character vector in modified snakecase
}
\description{
Essentially snakecase but preserving parenthesis
}
