% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_structural_metadata.R
\name{update_structural_metadata}
\alias{update_structural_metadata}
\title{Update structural metadata}
\usage{
update_structural_metadata(
  data,
  metadata,
  primary_key = "",
  foreign_key = "",
  additional_elements = tibble::tibble()
)
}
\arguments{
\item{data}{Any named object. Expects a table but will work
superficially with lists or named vectors.}

\item{metadata}{Data frame. Output from \code{create_structural_metadata}}

\item{primary_key}{Character. OPTIONAL Primary key in the data}

\item{foreign_key}{Character. OPTIONAL Foreign key or keys in the data}

\item{additional_elements}{data frame. OPTIONAL Empty tibble with structural
metadata elements and their types.}
}
\value{
data.frame
}
\description{
Appends rows and/or columns to existing metadata, change primary key and/or
adds foreign keys.
}
