% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_rds.R
\name{write_rds}
\alias{write_rds}
\title{Write data RDS}
\usage{
write_rds(r_obj, rds_file)
}
\arguments{
\item{r_obj}{R object to be saved. Likely a data frame}

\item{rds_file}{String. Path to rds file in the repo}
}
\value{
Path to rds file
}
\description{
This function will write an RDS file for an R object if the file
\enumerate{
\item doesnt exist or 2) the object or file have changed.
}
}
\details{
It is useful when working with data streams (airtable, googledrive, dropbox,
etc) that depend on account based access - particularly through an
institution or via an employer. If that account loses access to the data
stream then an entire pipeline may go down. Having a copy of the data
stored as RDS allows you to modify your pipeline to use the local copy
of the data instead.
}
\examples{
 # pretend we got this from a work associated dropbox folder
 df <- data.frame(a = 1:10)
 # write rds file somewhere
 rds_path <- write_rds(df, "example.rds")
 # read it back in
 df_from_rds < readRDS(rds_path)-
}
