% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_free_text_log.R
\name{create_free_text_log}
\alias{create_free_text_log}
\title{Create Translation Log}
\usage{
create_free_text_log(response_data, form_schema, url, type_to_keep = "text")
}
\arguments{
\item{response_data}{data.frame of ODK questionnaire responses}

\item{form_schema}{data.frame or flattened ODK form schema}

\item{url}{The ODK submission URL excluding the uuid identifier}

\item{type_to_keep}{String. Regex pattern passed to \code{stringr::str_detect}.}
}
\value{
data.frame validation log
}
\description{
Collates free text responses from text fields in the survey
data. Some language detection is performed and placed in the log notes section
for possible translation.
}
\examples{
\dontrun{

odk_schema_data <- schema_from_odk_xlsx_template(file_path = "inst/example_odk_schema.xlsx")

create_free_text_log(response_data = semi_clean_data,
                      form_schema = odk_schema_data,
                      url = "https://odk.xyz.io/#/projects/project-name/submissions")
}

}
