% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_freetext_log.R
\name{create_freetext_log}
\alias{create_freetext_log}
\title{Create Free Text Log}
\usage{
create_freetext_log(response_data, form_schema, url, ...)
}
\arguments{
\item{response_data}{data.frame ODK questionnaire response data}

\item{form_schema}{data.frame ODK flattened form schema data}

\item{url}{The ODK submission URL excluding the uuid identifier}

\item{...}{other arguments passed to `othertext_lookup()`}
}
\value{
data.frame validation log
}
\description{
Creates custom validation log for 'other: explain' free text responses that may contain valid
multi-choice options.
}
