% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_translation_log.R
\name{create_translation_log}
\alias{create_translation_log}
\title{Create Translation Log}
\usage{
create_translation_log(response_data, form_schema, url)
}
\arguments{
\item{response_data}{data.frame of ODK questionnaire responses}

\item{form_schema}{data.frame od flattened ODK form schema}

\item{url}{The ODK submission URL excluding the uuid identifier}
}
\value{
data.frame validation log
}
\description{
Collates free text responses from 'other' and 'notes' fields in the survey
data. Some language detection is performed, and those records that are not
detected as english are returned in a validation log for possible translation.
}
