% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_language.R
\name{detect_language}
\alias{detect_language}
\title{Detect Language}
\usage{
detect_language(text)
}
\arguments{
\item{text}{character any text string}
}
\value{
character estimate for language abbreviation
}
\description{
A function that extracts the top guess of the language of a piece of text.
}
\details{
Utilizes the stringi package encoding detector as the means to infer language.
}
\examples{
detect_language(text = "buongiorno")

}
\seealso{
\code{\link[stringi:stri_enc_detect]{stringi::stri_enc_detect()}}
}
