% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_googledrive_files.R
\name{download_googledrive_files}
\alias{download_googledrive_files}
\title{Download Google Drive Files}
\usage{
download_googledrive_files(
  key_path,
  drive_path,
  search_pattern,
  MIME_type = NULL,
  out_path
)
}
\arguments{
\item{key_path}{character path to Google authentication key}

\item{drive_path}{character The Google drive folder path}

\item{search_pattern}{character A search pattern for files in the Google drive}

\item{MIME_type}{character Google Drive file type, file extension, or MIME type.}

\item{out_path}{character The local file directory for files to be downloaded to}
}
\value{
a character vector of files downloaded
}
\description{
For a given Google Drive folder this function will find and download all files
matching a given pattern.
}
\details{
Note: This relies on the \code{googledrive::drive_ls()} function which uses a search function
and is not deterministic when recursively searching. Please pay attention to what is returned.
}
\examples{
\dontrun{
  download_googledrive_files(
  key_path = here::here("./key.json"),
  drive_path = "https://drive.google.com/drive/u/0/folders/asdjfnasiffas8ef7y7y89rf",
  search_pattern = ".*\\\\.xlsx",
  out_path = here::here("data/project_data/")
  )
}

}
\seealso{
\code{\link[googledrive:drive_ls]{googledrive::drive_ls()}}
}
