% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropbox_upload.R
\name{dropbox_upload}
\alias{dropbox_upload}
\title{Dropbox Upload}
\usage{
dropbox_upload(log, file_path, dropbox_path)
}
\arguments{
\item{file}{path to local file for upload}

\item{path}{relative dropbox path}
}
\value{
performs drop box upload
}
\description{
Upload a local file to dropbox and handle authentication
}
