% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropbox_upload.R
\name{dropbox_upload}
\alias{dropbox_upload}
\title{Dropbox Upload}
\usage{
dropbox_upload(log, file_path, dropbox_path, compress = TRUE)
}
\arguments{
\item{log}{dataframe. Validation Log for OH cleaning pipelines. Will work with any tabular data.}

\item{file_path}{character. local file path for upload}

\item{dropbox_path}{character. relative dropbox path}

\item{compress}{logical. Should files over 300mb be compressed?}
}
\value{
performs drop box upload
}
\description{
Upload a local file to dropbox and handle authentication.
}
\details{
This is a wrapper of \code{rdrop2::drop_upload()} which first reads in a local
CSV file and then uploads to a DropBox path.
}
\examples{
\dontrun{
    dropbox_upload(
    kzn_animal_ship_semiclean,
    file_path = here::here("outputs/data.csv"),
    dropbox_path = "XYZ/Data/semi_clean_data"
    )
}

}
