% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dropbox_val_logs.R
\name{get_dropbox_val_logs}
\alias{get_dropbox_val_logs}
\title{Get Dropbox Validation Logs}
\usage{
get_dropbox_val_logs(file_name, folder, path_name)
}
\arguments{
\item{file_name}{character file name with extension of the validation log}

\item{folder}{character the folder the log is saved in on drop box. Can be NULL if not in subfolder.}

\item{path_name}{character the default drop box path}
}
\value{
tibble a Validation Log
}
\description{
Downloads existing validation logs that are stored on dropbox
}
\details{
This function will check if the log exists and return NULL if not. Else it will
locally download the file to 'dropbox_validations' directory and read in to the
session.
}
\examples{
\dontrun{
 get_dropbox_val_logs(file_name = "log.csv", folder = NULL)
}

}
