% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_odk_form_schema.R
\name{get_odk_form_schema}
\alias{get_odk_form_schema}
\title{Get ODK Questionnaire Schema Info}
\usage{
get_odk_form_schema(
  url,
  un = Sys.getenv("ODK_USERNAME"),
  pw = Sys.getenv("ODK_PASSWORD"),
  odkc_version = Sys.getenv("ODKC_VERSION")
)
}
\arguments{
\item{url}{character The survey URL}

\item{un}{character The ODK account username}

\item{pw}{character The ODK account password}

\item{odkc_version}{character The ODKC Version string}
}
\value{
data frame of survey responses
}
\description{
This function handles the authentication and pulling of questionnaire
form schema information.
}
\details{
This is a wrapper around the \code{ruODK} package. It handles the setup and
authentication. See \url{https://github.com/ropensci/ruODK}
}
\examples{
\dontrun{
    get_odk_form_schema(url ="https://odk.xyz.io/v1/projects/5/forms/survey.svc",
    un = Sys.getenv("ODK_USERNAME"),
    pw = Sys.getenv("ODK_PASSWORD"),
    odkc_version = Sys.getenv("ODKC_VERSION"))
}

}
\seealso{
\code{\link[ruODK:form_schema_ext]{ruODK::form_schema_ext()}}
}
