% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_odk_responses.R
\name{get_odk_responses}
\alias{get_odk_responses}
\title{Get ODK Questionnaire Response Data}
\usage{
get_odk_responses(
  url,
  un = Sys.getenv("ODK_USERNAME"),
  pw = Sys.getenv("ODK_PASSWORD")
)
}
\arguments{
\item{url}{character The survey URL}

\item{un}{character The ODK account username}

\item{pw}{character The ODK account password}
}
\value{
data.frame of flattened survey responses
}
\description{
This function handles the authentication and pulling of submission
data for ODK Questionnaires.
}
