% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_report_urls.R
\name{make_report_urls}
\alias{make_report_urls}
\title{Make the urls for the reports}
\usage{
make_report_urls(aws_deploy_target, pattern = "")
}
\arguments{
\item{aws_deploy_target}{List. Output from aws_s3_upload}

\item{pattern}{String. Regex pattern for matching file paths}
}
\description{
Make the urls for the reports
}
\author{
Collin Schwantes
}
