% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obfuscate_gps.R
\name{obfuscate_gps}
\alias{obfuscate_gps}
\alias{obfuscate_lat}
\alias{obfuscate_lon}
\title{Obfuscate GPS}
\usage{
obfuscate_gps(
  x,
  precision = 2,
  fuzz = 0.125,
  type = c("lat", "lon"),
  func = min,
  ...
)

obfuscate_lat(x, precision = 2, fuzz = 0.125)

obfuscate_lon(x, precision = 2, fuzz = 0.125)
}
\arguments{
\item{x}{Numeric. Vector of gps points}

\item{precision}{Integer. Number of digits to keep. See \code{round} for more
details}

\item{fuzz}{Numeric. Positive number indicating how much error to introduce
to the gps measurements. This is used to generate the random uniform
distribution \code{runif(1,min = -fuzz, max = fuzz)}}

\item{type}{Character. One of "lat" or "lon"}

\item{func}{Function. Function used in \code{get_precision}}

\item{...}{Additional arguments for func.}
}
\value{
Numeric. A vector of fuzzed and rounded GPS points

Numeric vector

Numeric vector
}
\description{
This function fuzzes gps points by first adding
error then rounding to a certain number of digits.
}
\examples{

# make data
gps_data  <- data.frame(lat = c(1.0001, 10.22223, 4.00588),
                        lon = c(2.39595, 4.506930, -60.09999901))

# Default obfuscation settings correspont to roughly a 27 by 27 km area
gps_data$lat |>
  obfuscate_gps(type = "lat")

# Obfuscation can be made more or less precise by changing the number of
# decimal points included or modifying the amount of fuzz (error)
# introduced
gps_data$lon |>
  obfuscate_gps(precision = 4, fuzz = 0.002, type = "lon")

### working at the poles
gps_data_poles  <- data.frame(lat = c(89.0001, 89.22223, -89.8881),
                              lon = c(2.39595, 4.506930, -60.09999901))


gps_data_poles$lat |>
  obfuscate_gps(fuzz = 1, type = "lat")


### working at the 180th meridian
gps_data_180  <- data.frame(lat = c(2, 3, 4),
                            lon = c(179.39595, -179.506930, -178.09999901))
gps_data_180$lon |>
  obfuscate_gps(fuzz = 1, type = "lon")

### GPS is on the fritz!
\dontrun{
gps_data_fritz <- data.frame(lat = c(91, -91, 90),
                             lon = c(181.0001, -181.9877, -178.09999901))
gps_data_fritz$lon |>
  obfuscate_gps(fuzz = 1, type = "lon")

gps_data_fritz$lat |>
  obfuscate_gps(fuzz = 1, type = "lat")
}

}
