% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohcleandat-package.R
\docType{package}
\name{ohcleandat-package}
\alias{ohcleandat}
\alias{ohcleandat-package}
\title{ohcleandat: One Health Data Cleaning and Quality Checking Package}
\description{
This package provides useful functions to orchestrate analytics and data cleaning pipelines for One Health projects.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://One-Health-Research-Consulting.github.io/ohcleandat/}
}

}
\author{
\strong{Maintainer}: Collin Schwantes \email{cjschwantesll@gmail.com} (\href{https://orcid.org/0000-0003-4014-4896}{ORCID})

Authors:
\itemize{
  \item Johana Teigen \email{joteigen14@gmail.com} (\href{https://orcid.org/0000-0002-6209-2321}{ORCID})
  \item Ernest Guevarra \email{ernest@guevarra.io} (\href{https://orcid.org/0000-0002-4887-4415}{ORCID})
  \item Dean Marchiori \email{dean@wavedatalabs.com.au} (\href{https://orcid.org/0000-0002-3430-7225}{ORCID})
  \item Melinda Rostal \email{rostal@onehealthresearch.net} (\href{https://orcid.org/0000-0002-6563-5280}{ORCID})
}

Other contributors:
\itemize{
  \item EcoHealth Alliance (https://ror.org/02zv3m156) [copyright holder, funder]
}

}
\keyword{internal}
