% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/othertext_lookup.R
\name{othertext_lookup_from_odk_excel}
\alias{othertext_lookup_from_odk_excel}
\title{Create other text lookup from ODK excel template}
\usage{
othertext_lookup_from_odk_excel(file_path)
}
\arguments{
\item{file_path}{String. Path to excel template file}
}
\value{
tibble. Tibble with fields name and other name
}
\description{
Create other text lookup from ODK excel template
}
