% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_excel_all_sheets.R
\name{read_excel_all_sheets}
\alias{read_excel_all_sheets}
\title{Reads all tabs from an excel workbook}
\usage{
read_excel_all_sheets(
  file,
  add_primary_key_field = FALSE,
  primary_key = "primary_key"
)
}
\arguments{
\item{file}{character. File path to an excel file}

\item{add_primary_key_field}{Logical. Should a primary key field be added?}

\item{primary_key}{character. The column name for the unique identifier to be added to the data.}
}
\value{
list
}
\description{
For a given excel file, this will detect all sheets, and iteratively read
all sheets and place them in a list.

If primary keys are added, the primary key is the triplet of the file,
sheet name, and row number e.g. "file_xlsx_sheet1_1". Row numbering is based
on the data ingested into R. R automatically skips empty rows at the beginning
of the spreadsheet so id 1 in the primary key will belong to the first row
with data.
}
\note{
The primary key method is possible because Excel forces sheet names
 to be unique.
}
\examples{
 \dontrun{
# Adding primary key field
read_excel_all_sheet(file = "test_pk.xlsx",add_primary_key_field = TRUE)

# Don't add primary key field
read_excel_all_sheet(file = "test_pk.xlsx")

    }

}
