% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_deletions.R
\name{remove_deletions}
\alias{remove_deletions}
\title{Utility function to identify records for deletion}
\usage{
remove_deletions(x, val = "Delete")
}
\arguments{
\item{x}{input vector}

\item{val}{The value to check for inequality. Defaults to 'Delete'}
}
\value{
logical vector
}
\description{
To be used within dplyr::filter. The function returns a logical vector
with TRUE resulting from values that are not equal to the `val` argument. Also
protects from na values.
}
\details{
Used within verbs such as `all_of()` this can work effectively across all
columns in a data frame. See examples
}
\examples{
data |> filter(if_all(everything(), remove_deletions))
}
