% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_checks.R
\name{validation_checks}
\alias{validation_checks}
\title{Validation Correction Checks}
\usage{
validation_checks(validation_log, before_data, after_data, idcol)
}
\arguments{
\item{validation_log}{tibble Validation log}

\item{before_data}{tibble Data before corrections}

\item{after_data}{tibble Data after corrections}

\item{idcol}{character the primary key for the 'after_data'}
}
\value{
NULL if passed or stops with error
}
\description{
Runs some checks on data before and after validation to test expectations
}
